+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to the **SURF** project will be documented in this file.
 The format is based on [Keep a Changelog][kacl], and this project adheres to
 [Semantic Versioning][sv].
 
+## 0.0.0.9001 - XXXX-XX-XX
+
+### Added
+- new function `rangeNorm` and associated doc page
+- new function `calcMode` and associated doc page
+- new function `f2n` and associated doc page
+- doc file for `testImps.RData` example data
+
+### Change
+- included licensing header in all source files
+
 ## 0.0.0.9000 - 2017-11-03
 
 ### Added